Fix version date calculation in corner cases (#149)
Previously, if the latest published stub version was higher
than the upstream version (which could happen with our old
versioning scheme), we always used the current date as
last component. This failed when uploading the same package
twice in a day – but also in our stub_uploader tests, which
simulate a new upload.